Generic input reader
